# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Seat Reservation API
  version: 1.0.0
  description: |
    Sells assigned seats for an event and never double-sells, over-allocates a user, or double-charges a retried request.

    ## Quick start
    1. `POST /auth/token` with `{"user_id":"alice"}` and copy `token`. For an admin token also send `admin_secret`.
    2. Click **Authorize** (top right) and paste the token. Do not type the word `Bearer`.
    3. As admin: `POST /shows`. As any user: `reserve` (confirmed immediately) or `hold` then `confirm`. Use a **new `idempotency_key` per order** and reuse it only to retry that same order.

    ## Conventions
    * Money is an **integer number of paise**. Fractional values are rejected with 400.
    * Identity is the token subject. A `user_id` in a request body is ignored.
    * Multi-seat requests are **all-or-nothing**: if any seat is unavailable nothing is taken.
    * Business declines are `409` with `{"error": "<code>", "message": "..."}`. Domain outcomes never return 5xx.
    * Every response carries `X-Request-ID` (send your own to correlate logs).

    ## Error codes
    | HTTP | `error` | Meaning |
    |---|---|---|
    | 400 | `invalid_request` | malformed JSON, missing/invalid field, duplicate seats, bad idempotency key |
    | 400 | `invalid_seat` | seat does not exist in this show |
    | 401 | `unauthorized` | missing, malformed, expired or wrongly signed token |
...
    | 404 | `show_not_found`, `reservation_not_found` | unknown id |
...
servers:
...
paths:
...
  /ops/logs:
    get:
      tags: [Ops]
      operationId: recentLogs
      security: []
      summary: Recent structured log events
      description: |
        The most recent request-access and business events (newest last), read from an in-memory buffer of 2000 events.
        Safe to expose: only an allow-list of fields is returned (no headers, request bodies, tokens, secrets or stack traces),
        string values are scrubbed of JWT/bearer/connection-URL patterns, and framework logs are never captured.
        Every entry carries the `request_id` that is also returned in the `X-Request-ID` response header.
        Disabled (404) when `PUBLIC_LOGS=false`.
      parameters:
        - name: limit
          in: query
          description: Maximum number of entries (1-1000)
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
        - name: request_id
          in: query
          description: Only entries for this request id
          schema: { type: string }
      responses:
        '200':
          description: Log entries
          content:
            application/json:
              example:
                count: 1
                buffer_capacity: 2000
                entries:
                  - { timestamp: "2026-10-03T12:34:39.437Z", level: INFO, logger: access, message: request, request_id: 4ffff278-52da-4869-8117-6e676926ea29, method: POST, path: /shows/9973c587/reserve, status: 201, duration_ms: 2, user_id: alice }
        '404':
          description: Disabled
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token from `POST /auth/token`
...
  schemas:
...
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }